Moving to definitions instead of definitions_api
