+++ exhibitera/apps/Exhibitera_Apps.py
@@ -29,7 +29,7 @@ import helper_utilities
 
 # Api Modules 
 from api.system_api import system_api
-from api.definitions_api import definitions_api
+from api.definitions import definitions
 # If we're not on Linux, prepare to use the webview
 if sys.platform != 'linux':
     import webview
@@ -125,7 +125,7 @@ app.mount("/thumbnails",
           name="thumbnails")
 
 app.include_router(system_api.router)
-app.include_router(definitions_api.router)
+app.include_router(definitions.router)
 @lru_cache()
 def get_config():
     return const_config
